Use user defined hover text for image in draggable
With changing the image hover handling when metadata were
introduced, images within draggables used the images
metadata title for the hover text, but not what the user
specified in the editor of Drag-and-Drop.

This change overrides the image hover text for draggable
images. The hover for both, the draggable container and
the image itself will now be the same.