Use user defined hover text for static images
With changing the image hover handling when metadata were
introduced, static images used the images
metadata title for the hover text, but not what the user
specified in the editor of Drag-and-Drop.

This change overrides the image hover text for static
images. It will be what's specified by the user, so
possibly empty and not displayed at all.